LibGfx: Move (de)coders to a dedicated `Formats` directory
