+++ Userland/Libraries/LibGfx/Formats/BMP/BMPLoader.h
