+++ Userland/Libraries/LibGfx/Formats/BMP/BMPWriter.h
