+++ Userland/Libraries/LibGfx/Formats/DDS/DDSLoader.h
