+++ Userland/Libraries/LibGfx/Formats/GIF/GIFLoader.cpp
