+++ Userland/Libraries/LibGfx/Formats/GIF/GIFLoader.h
