+++ Userland/Libraries/LibGfx/Formats/ICO/ICOLoader.cpp
