+++ Userland/Libraries/LibGfx/Formats/ICO/ICOLoader.h
