+++ Userland/Libraries/LibGfx/Formats/JPG/JPGLoader.cpp
