+++ Userland/Libraries/LibGfx/Formats/JPG/JPGLoader.h
