+++ Userland/Libraries/LibGfx/Formats/PGM/PGMLoader.h
