+++ Userland/Libraries/LibGfx/Formats/PNG/PNGLoader.cpp
