+++ Userland/Libraries/LibGfx/Formats/PNG/PNGLoader.h
