+++ Userland/Libraries/LibGfx/Formats/PNG/PNGWriter.cpp
