+++ Userland/Libraries/LibGfx/Formats/PNG/PNGWriter.h
