+++ Userland/Libraries/LibGfx/Formats/PPM/PPMLoader.cpp
@@ -5,12 +5,12 @@
  */
 
 #include "PPMLoader.h"
-#include "PortableImageLoaderCommon.h"
-#include "Streamer.h"
 #include <AK/Endian.h>
 #include <AK/LexicalPath.h>
 #include <AK/ScopeGuard.h>
 #include <AK/StringBuilder.h>
+#include <LibGfx/PortableImageLoaderCommon.h>
+#include <LibGfx/Streamer.h>
 #include <string.h>
 
 namespace Gfx {
