+++ Userland/Libraries/LibGfx/Formats/QOI/QOILoader.h
